# perf: speed up time loop in processed var init sens