# Description

Speeds up initial sensitivitiy calculation in ProcessedVariable. The diagcat and vertcat is done once instead of every timestep

I did a simple test using the DFN model and one input paramter (current), before the change the timings were (over 10 runs)

```
(env-pybammsolvers) ➜  PyBaMM git:(develop) ✗ python test.py
Time taken to compute sensitivities: 0.05452 seconds
Time taken to compute sensitivities: 0.04619 seconds
Time taken to compute sensitivities: 0.04333 seconds
Time taken to compute sensitivities: 0.04116 seconds
Time taken to compute sensitivities: 0.04146 seconds
Time taken to compute sensitivities: 0.04142 seconds
Time taken to compute sensitivities: 0.04156 seconds
Time taken to compute sensitivities: 0.04135 seconds
Time taken to compute sensitivities: 0.04133 seconds
Time taken to compute sensitivities: 0.04186 seconds
```

After the change the timings were:

```
(env-pybammsolvers) ➜  PyBaMM git:(develop) ✗ python test.py
Time taken to compute sensitivities: 0.02355 seconds
Time taken to compute sensitivities: 0.02262 seconds
Time taken to compute sensitivities: 0.02266 seconds
Time taken to compute sensitivities: 0.02286 seconds
Time taken to compute sensitivities: 0.02309 seconds
Time taken to compute sensitivities: 0.02277 seconds
Time taken to compute sensitivities: 0.02158 seconds
Time taken to compute sensitivities: 0.01918 seconds
Time taken to compute sensitivities: 0.01988 seconds
Time taken to compute sensitivities: 0.02149 seconds
```

## Type of change



# Important checks:

Please confirm the following before marking the PR as ready for review:
- No style issues: `nox -s pre-commit`
- All tests pass: `nox -s tests`
- The documentation builds: `nox -s doctests`
- Code is commented for hard-to-understand areas
- Tests added that prove fix is effective or that feature works